feat(server): adding event emitter
