+++ server/app/facade.go
@@ -4,7 +4,13 @@ import (
 	"context"
 
 	"github.com/kubeshop/tracetest/server/executor"
+	"github.com/kubeshop/tracetest/server/executor/event"
+	"github.com/kubeshop/tracetest/server/executor/pollingprofile"
+	"github.com/kubeshop/tracetest/server/executor/trigger"
 	"github.com/kubeshop/tracetest/server/model"
+	"github.com/kubeshop/tracetest/server/subscription"
+	"github.com/kubeshop/tracetest/server/tracedb"
+	"go.opentelemetry.io/otel/trace"
 )
 
 type runnerFacade struct {
@@ -24,4 +30,70 @@ func (rf runnerFacade) RunTransaction(ctx context.Context, tr model.Transaction,
 
 func (rf runnerFacade) RunAssertions(ctx context.Context, request executor.AssertionRequest) {
 	rf.assertionRunner.RunAssertions(ctx, request)
+}
+
+func newRunnerFacades(
+	ppRepo *pollingprofile.Repository,
+	testDB model.Repository,
+	appTracer trace.Tracer,
+	tracer trace.Tracer,
+	subscriptionManager *subscription.Manager,
+	triggerRegistry *trigger.Registry,
+) *runnerFacade {
+	eventEmitter := event.NewEmitter(testDB, subscriptionManager)
+
+	execTestUpdater := (executor.CompositeUpdater{}).
+		Add(executor.NewDBUpdater(testDB)).
+		Add(executor.NewSubscriptionUpdater(subscriptionManager))
+
+	assertionRunner := executor.NewAssertionRunner(
+		execTestUpdater,
+		executor.NewAssertionExecutor(tracer),
+		executor.InstrumentedOutputProcessor(tracer),
+		subscriptionManager,
+		eventEmitter,
+	)
+
+	pollerExecutor := executor.NewPollerExecutor(
+		ppRepo,
+		tracer,
+		execTestUpdater,
+		tracedb.Factory(testDB),
+		testDB,
+		eventEmitter,
+	)
+
+	tracePoller := executor.NewTracePoller(
+		pollerExecutor,
+		ppRepo,
+		execTestUpdater,
+		assertionRunner,
+		subscriptionManager,
+		eventEmitter,
+	)
+
+	runner := executor.NewPersistentRunner(
+		triggerRegistry,
+		testDB,
+		execTestUpdater,
+		tracePoller,
+		tracer,
+		subscriptionManager,
+		tracedb.Factory(testDB),
+		testDB,
+		eventEmitter,
+	)
+
+	transactionRunner := executor.NewTransactionRunner(
+		runner,
+		testDB,
+		subscriptionManager,
+	)
+
+	return &runnerFacade{
+		runner:            runner,
+		transactionRunner: transactionRunner,
+		assertionRunner:   assertionRunner,
+		tracePoller:       tracePoller,
+	}
 }
